==> codes/0_Index.Rmd <==
---
title: "Data analysis of the GH002"
author: "Ran An"
date: "`r Sys.Date()`"
output: bookdown::gitbook
site: bookdown::bookdown_site
---

```{r child = '1_create_phyloseq_object.Rmd'}
```

```{r child = '1_sample_numbers_per_group.Rmd'}
```

```{r child = '2_QC_quality_control.Rmd'}
```

```{r child = '3_data_exploration.Rmd'}
```

```{r child = '4.1_young_vs_elderly_alpha_diversity.Rmd'}
```

```{r child = '4.2_young_vs_elderly_beta_diversity.Rmd'}
```

```{r child = '4.3.significantly.different.taxa.Rmd'}
```

```{r child = '5.1_intervention_alpha_div.Rmd'}
```

```{r child = '5.2_intervention_beta_div.Rmd'}
```

```{r child = '5.3_dif_taxa_post_intervention_y_vs_E.Rmd'}
```
