==> codes/4.2_young_vs_elderly_beta_diversity.Rmd <==
---
title: "young vs elderly-beta diversity"
author: "Ran An"
date: "`r Sys.Date()`"
output: bookdown::gitbook
site: bookdown::bookdown_site
---
##Baseline beta diversity
```{r, warning=FALSE, message=FALSE, echo=FALSE}
library(microbiome) # data analysis and visualisation
library(phyloseq) # also the basis of data object. Data analysis and visualisation
library(microbiomeutilities) # some utility tools 
library(RColorBrewer) # nice color options
library(ggpubr) # publication quality figures, based on ggplot2
library(DT) # interactive tables in html and markdown
library(data.table) # alternative to data.frame
library(dplyr) # data handling  
library(vegan)
library(cowplot)
```

Baseline  data preparation
```{r}
ps1_baseline <- readRDS("./phyobjects/ps_t0.rds") #read in the baseline phyloseq object
```

Relative.abundance based beta diversity analysis -MDS
```{r}
ps1.rel <- transform(ps1_baseline, "relative.abundance")
dist.rel.uf <- phyloseq::distance(ps1.rel, method = "unifrac") ## Unifrac
dist.rel.wuf <- phyloseq::distance(ps1.rel, method = "wunifrac") ## Weighted Unifrac
dist.rel.bc <- distance(ps1.rel, method = "bray") ## Bray-Curtis

#MDS + Weighted Unifrac
ord_rel.wuf <- ordinate(ps1.rel, method = "MDS", distance = dist.rel.wuf)
p <- plot_ordination(ps1.rel, ord_rel.wuf, shape = "Group")
p <- p + scale_shape_manual(values= c(19,24,15,3,7,8,16,12))
p <- p + theme_bw() + ggtitle("PCoA + Weighted Unifrac") ## add title and plain background


#MDS + unWeighted Unifrac
ord_rel.uf <- ordinate(ps1.rel, method = "MDS", distance = dist.rel.uf)
p1 <- plot_ordination(ps1.rel, ord_rel.uf, shape = "Group")
p1 <- p1 + scale_shape_manual(values= c(19,24,15,3,7,8,16,12))
p1 <- p1 + theme_bw() + ggtitle("PCoA + unWeighted Unifrac") ## add title and plain background

fig<-plot_grid(p, p1,labels=c("A", "B"), ncol = 2, nrow = 1)
fig
```
