==> codes/4.3.significantly.different.taxa.Rmd <==
---
title: "4.4.significantly.different.taxa"
author: "Ran An"
date: "`r Sys.Date()`"
output: bookdown::gitbook
site: bookdown::bookdown_site
---
## Significantly.different.taxa
```{r,warning=FALSE,message=FALSE, echo=FALSE}
library(ggplot2)
library(ggpubr) # add p value
library(dplyr)
library(tibble) # label outliers
library(cowplot)#combine plots
```


Load data into the program
```{r}
#load the relative abundance data
L6 <- read.csv("./input files/biom_GH002_L6.csv", check.names = FALSE, row.names = 1,stringsAsFactors = FALSE,header = TRUE)
L6 <- t(L6)

#load mapping file
meta_p1 <-read.csv("./input files/mapping_GH002.csv",check.names=FALSE, row.names = 1,stringsAsFactors = FALSE, header = TRUE)

# merge the mapping file and L6 data
p6 <- merge(meta_p1,L6, by.x='OTU ID')
colnames(p6)[1:6] # check the quality
p6[] <- lapply(p6, gsub, pattern = "NA", replacement = "0", fixed = TRUE) #change NA to "0"

p6 <- subset(p6, Intervention != "MOCK")   #get rid of mocks
nrow(p6)#check if subset succeed
```


Get the baseline data, in order to look into age group effects
```{r}
baseline<-subset(p6, Time == "pre")
G1<- rownames(subset(baseline, Group=="Young"))
G2<- rownames(subset(baseline, Group=="Elderly"))

baseline <-t(baseline)
baseline<-baseline[48:271,]
```


```{r, echo=FALSE}
#function for comparison
comparison <- function (dat, G1, G2, p.adjust.method = "BH", sort = FALSE, 
                        paired = FALSE) 
{
  samples <- colnames(dat)
  levels <- rownames(dat)
  M <- matrix(data = NA, length(levels), 1)
  rownames(M) <- levels
  for (i in 1:length(levels)) {
    lvl <- levels[i]
    l.g1 <- dat[lvl, G1]
    l.g2 <- dat[lvl, G2]
    p <- wilcox.test(as.numeric(l.g1), as.numeric(l.g2), 
                     paired = FALSE)$p.value # unpaired for young vs elderly
    M[i, 1] <- p
  }
  if (!is.null(p.adjust.method)) {
    cor.p <- p.adjust(M, method = p.adjust.method)
    names(cor.p) <- rownames(M)
  }
  else {
    cor.p <- as.vector(M)
    names(cor.p) <- rownames(M)
  }
  if (sort) {
    cor.p <- sort(cor.p)
  }
  cor.p
}

result<-comparison(baseline, G1, G2)
df<-as.data.frame(result)
#df

#write.csv(df, "./tables/different_taxa_wilcox_y_vs_e.csv")
```


Plot six genera which have been found significantly different between the microbiota of young adults and elderly
```{r}
# convert factor to numeric
baseline <- subset(p6, Time == "pre")
nrow(baseline) # check whether code above worked

# convert factor to numeric
baseline_1 <- baseline[,c(1:47)]
baseline_2 <- baseline[,c(48:271)]
indx <- sapply(baseline_2, is.character)
baseline_2[indx] <- lapply(baseline_2[indx], function(x) as.numeric(as.character(x)))
baseline<-cbind(baseline_1,baseline_2)
```


Combine multiple plots
```{r}
p <- ggplot(baseline, aes(x=Group, y=k__Bacteria_p__Firmicutes_c__Clostridia_o__Clostridiales_f__Lachnospiraceae_g__LachnospiraceaeUCG_008, shape = Group))
p <- p + geom_boxplot(outlier.shape=NA)
p <- p + geom_jitter(position=position_jitter(0.2), cex=1.5) + scale_shape_manual(values=c(16,2))
p <- p + labs(y="Relative abundance")
p <- p + labs(x="")
p <- p + theme_classic()
p <- p + theme(legend.position = "none")
p <- p + scale_x_discrete(limits=c("Young","Elderly"))
p <- p + ggtitle("Enterorhabdus") + theme(plot.title = element_text(size = 12))


p1 <- ggplot(baseline, aes(x=Group, y=k__Bacteria_p__Actinobacteria_c__Coriobacteriia_o__Coriobacteriales_f__Coriobacteriaceae_g__uncultured, shape = Group))
p1 <- p1 + geom_boxplot(outlier.shape=NA)
p1 <- p1 + geom_jitter(position=position_jitter(0.2), cex=1.5)+ scale_shape_manual(values=c(16,2))
p1 <- p1 + labs(y="Relative abundance",size=0.4)
p1 <- p1 + labs(x="")
p1 <- p1 + theme_classic()
p1 <- p1 + theme(legend.position = "none")
p1 <- p1 + scale_x_discrete(limits=c("Young","Elderly"))
p1 <- p1 + ggtitle("Coriobacteriaceae;g__uncultured") + theme(plot.title = element_text(size = 11))


p2 <- ggplot(baseline, aes(x=Group, y=k__Bacteria_p__Firmicutes_c__Clostridia_o__Clostridiales_f__FamilyXIII_g__Mogibacterium, shape = Group))
p2 <- p2 + geom_boxplot(outlier.shape=NA)
p2 <- p2 + geom_jitter(position=position_jitter(0.2), cex=1.5)+ scale_shape_manual(values=c(16,2))
p2 <- p2 + labs(y="Relative abundance")
p2 <- p2 + labs(x="")
p2 <- p2 + theme_classic()
p2 <- p2 + theme(legend.position = "none")
p2 <- p2 + scale_x_discrete(limits=c("Young","Elderly"))
p2 <- p2 + ggtitle("Mogibacterium") + theme(plot.title = element_text(size = 12))


p3 <- ggplot(baseline, aes(x=Group, y=k__Bacteria_p__Firmicutes_c__Clostridia_o__Clostridiales_f__Lachnospiraceae_g__LachnospiraceaeUCG_008, shape = Group))
p3 <- p3 + geom_boxplot(outlier.shape=NA)
p3 <- p3 + geom_jitter(position=position_jitter(0.2), cex=1.5)+ scale_shape_manual(values=c(16,2))
p3 <- p3 + labs(y="Relative abundance")
p3 <- p3 + labs(x="")
p3 <- p3 + theme_classic()
p3 <- p3 + theme(legend.position = "none")
p3 <- p3 + scale_x_discrete(limits=c("Young","Elderly"))
p3 <- p3 + ggtitle("LachnospiraceaeUCG-008") + theme(plot.title = element_text(size = 12))


#k__Bacteria_p__Firmicutes_c__Clostridia_o__Clostridiales_f__Ruminococcaceae_g__Ruminiclostridium6
#Ruminiclostridium6
p4 <- ggplot(baseline, aes(x=Group, y=k__Bacteria_p__Firmicutes_c__Clostridia_o__Clostridiales_f__Ruminococcaceae_g__Ruminiclostridium6, shape = Group))
p4 <- p4 + geom_boxplot(outlier.shape=NA)
p4 <- p4 + geom_jitter(position=position_jitter(0.2), cex=1.5)+ scale_shape_manual(values=c(16,2))
p4 <- p4 + theme_classic()
p4 <- p4 + theme(legend.position = "none")
p4 <- p4 + labs(y="Relative abundance")
p4 <- p4 + labs(x="")
p4 <- p4 + scale_x_discrete(limits=c("Young","Elderly"))
p4 <- p4 + ggtitle("Ruminiclostridium6") + theme(plot.title = element_text(size = 12))

library(cowplot) # combine graphs in one
plot_grid(p, p1, p2, p3, p4, labels=c("A", "B","C","D", "E"), ncol = 3, nrow = 2)
```
